requirements.txt
pyyaml
pydantic
python-dotenv
requests
pandas
matplotlib
networkx==2.8.8
